Initial protocol lib support.

+++ src/main/java/cn/nukkit/block/BlockClay.java
@@ -16,12 +16,12 @@ public class BlockClay extends BlockSolid {
     }
 
     @Override
-    public double getHardness() {
-        return 0.6;
+    public float getHardness() {
+        return 0.6f;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 3;
     }
