+++ src/main/java/cn/nukkit/block/BlockCobblestone.java
@@ -15,12 +15,12 @@ public class BlockCobblestone extends BlockSolid {
     }
 
     @Override
-    public double getHardness() {
+    public float getHardness() {
         return 2;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 30;
     }
