+++ src/main/java/cn/nukkit/block/BlockStairsPurpur.java
@@ -11,12 +11,12 @@ public class BlockStairsPurpur extends BlockStairs {
     }
 
     @Override
-    public double getHardness() {
-        return 1.5;
+    public float getHardness() {
+        return 1.5f;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 30;
     }
