+++ src/main/java/cn/nukkit/block/BlockStairsQuartz.java
@@ -14,12 +14,12 @@ public class BlockStairsQuartz extends BlockStairs {
     }
 
     @Override
-    public double getHardness() {
-        return 0.8;
+    public float getHardness() {
+        return 0.8f;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 4;
     }
